ignore some py2 specific tests on py3
