+++ myhdl/test/conftest.py
@@ -1,3 +1,5 @@
+import sys
+
 import py
 import pytest
 
@@ -7,6 +9,9 @@ from myhdl.conversion._verify import _simulators
 xfail = pytest.mark.xfail
 
 all_sims = list(_simulators)
+
+if sys.version_info[0] > 2:
+    collect_ignore = ['conversion/toVerilog/test_not_supported_py2.py']
 
 def pytest_addoption(parser):
     parser.addoption("--sim", action="store", choices=all_sims,
