+++ myhdl/test/conversion/toVerilog/test_NotSupported.py
@@ -48,20 +48,6 @@ class TestNotSupported(unittest.TestCase):
             return logic
         self.check(g, z, a)
 
-    def testBackquote(self):
-        a = Signal(bool())
-        z = Signal(bool())
-        def g(z, a):
-            @instance
-            def logic():
-                while 1:
-                    yield a
-                    z.next = 1
-                    `a`
-            return logic
-        self.check(g, z, a)
-
-
     def testClass(self):
         a = Signal(bool())
         z = Signal(bool())
@@ -99,19 +85,6 @@ class TestNotSupported(unittest.TestCase):
                 while 1:
                     yield a
                     z.next = z / a
-            return logic
-        self.check(g, z, a)
-
-    def testExec(self):
-        a = Signal(bool())
-        z = Signal(bool())
-        def g(z, a):
-            @instance
-            def logic():
-                while 1:
-                    yield a
-                    z.next = 1
-                    exec "1 + 2" in globals , locals
             return logic
         self.check(g, z, a)
